modules/hooks-mode/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hooks-mode"
version = "1.0.0"
description = "Generic mode hooks for context injection and tool moderation"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
